requirements.txt
claude-agent-sdk>=0.1.48
pandas>=2.3
psycopg2-binary>=2.9
python-dotenv>=1.2
SQLAlchemy>=2.0
streamlit>=1.55
tabulate>=0.10
